runops/requirements-dev.txt

ansible-lint>=24.6.0
yamllint>=1.35.1
molecule>=24.2.0
molecule-plugins[docker]>=23.5.3
pytest>=8.2.0
pre-commit>=3.8.0
